[Ref] Silence various trivial Clang warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@11323 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/CListCtrl.h
@@ -28,7 +28,7 @@ public:
 	template<size_t numItems>
 	void SetHeaders(const Header (&header)[numItems])
 	{
-		for(int i = 0; i < numItems; i++)
+		for(size_t i = 0; i < numItems; i++)
 		{
 			InsertColumn(i, header[i].text, header[i].mask, Util::ScalePixels(header[i].width, m_hWnd));
 		}
